fixed convert from input browse

+++ apps/electron-vgm/src/graphql/models/level3.model.ts
@@ -1,5 +1,5 @@
 import 'reflect-metadata'
-import { ObjectType, Field, Int } from '@nestjs/graphql'
+import { ObjectType, Field, Int, Float } from '@nestjs/graphql'
 import { Level2 } from './level2.model'
 import { Level4 } from './level4.model'
 
@@ -62,7 +62,7 @@ export class Level3 {
   @Field((type) => String, { nullable: true })
   duration?: string
 
-  @Field((type) => Int, { nullable: true })
+  @Field((type) => Float, { nullable: true })
   size?: number
 
   @Field((type) => [Level2])
