+++ apps/electron-vgm/src/graphql/models/level5.model.ts
@@ -1,5 +1,5 @@
 import 'reflect-metadata'
-import { ObjectType, Field, Int } from '@nestjs/graphql'
+import { ObjectType, Field, Int, Float } from '@nestjs/graphql'
 import { Level4 } from './level4.model'
 import { Level6 } from './level6.model'
 
@@ -62,7 +62,7 @@ export class Level5 {
   @Field((type) => String, { nullable: true })
   duration?: string
 
-  @Field((type) => Int, { nullable: true })
+  @Field((type) => Float, { nullable: true })
   size?: number
 
   @Field((type) => [Level4])
