+++ apps/electron-vgm/src/graphql/models/level7.model.ts
@@ -1,5 +1,5 @@
 import 'reflect-metadata'
-import { ObjectType, Field, Int } from '@nestjs/graphql'
+import { ObjectType, Field, Int, Float } from '@nestjs/graphql'
 import { Level6 } from './level6.model'
 // import { Level8 } from './level8.model'
 
@@ -62,7 +62,7 @@ export class Level7 {
   @Field((type) => String, { nullable: true })
   duration?: string
 
-  @Field((type) => Int, { nullable: true })
+  @Field((type) => Float, { nullable: true })
   size?: number
 
   @Field((type) => [Level6])
